Use one multiline asm statement
This guarantees suquential ordering of each assembly statement which was
*not* guaranteed by the previous use of each assembly statement in a
separate 'asm' statement.

This caused a bug where gcc would zero out one of the registers after
the asm statement using it had finished and use it to set unused syscall
parameters to 0.

This bug was caught after I turned on optimizations.